Enhances UI and cart addition logic
Improves the user interface of the favorites sidebar and movie card components.

Updates the "Add to Cart" button in the movie card to indicate if the item is already in the cart and disables the button accordingly. Also, adds a visual update to the "Add to Cart" button in the favorites sidebar.